Move calc back to utilities
