+++ src/features/restaurants/Dashboard.tsx
@@ -12,6 +12,7 @@ import { useNavigate } from 'react-router-dom'
 import { APP_ROUTES } from '../../utilities/constants'
 import { Button } from 'react-bootstrap'
 import { getLifts } from '../lifts/liftsSlice' 
+import { getUserDistance } from '../../utilities/utilities'
 
 export const Dashboard = () => {
   const dispatch = useDispatch<AppDispatch>()
@@ -32,6 +33,8 @@ export const Dashboard = () => {
   useEffect(() => {
     if (status === 'idle') {
       dispatch(getRestaurants())
+      getUserDistance(Number(activeUser.location.lat), Number(activeUser.location.long), Number(restaurant.location.lat), Number(restaurant.location.long))
+
     }
   }, [status, dispatch])
