+++ src/utilities/utilities.ts
@@ -114,3 +114,29 @@ export const cleanLiftsData = (lifts: Lifts[]): CleanedLiftsState[] => {
     }
   })
 }
+
+
+export const getUserDistance = (
+  userLat: number,
+  userLong: number,
+  restLat: number,
+  restLong: number
+): string => {
+  let R = 6371 // Radius of the earth in km
+  let dLat = deg2rad(restLat - userLat) // deg2rad below
+  let dLon = deg2rad(restLong - userLong)
+  let a =
+    Math.sin(dLat / 2) * Math.sin(dLat / 2) +
+    Math.cos(deg2rad(userLat)) *
+      Math.cos(deg2rad(restLat)) *
+      Math.sin(dLon / 2) *
+      Math.sin(dLon / 2)
+  let c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a))
+  let d = R * c // Distance in km
+  let miles = d / 1.609
+  return miles.toFixed(2)
+}
+
+function deg2rad(deg: number) {
+  return deg * (Math.PI / 180)
+}
